refactor(theme): 移除注册样式表机制，只提供根据样式表生成主题的方法
