+++ src/index.ts
@@ -2,11 +2,10 @@ import { GeometryCfg } from './geometry/base';
 import { PathCfg } from './geometry/path';
 import { IInteractionContext } from './interface';
 
-// 注册主题
-import { registerStyleSheet, registerTheme } from './core';
+// 注册黑暗主题
+import { getThemeByStylesheet, registerTheme } from './core';
 import { antvDark } from './theme/style-sheet/dark';
-registerStyleSheet('dark', antvDark);
-registerTheme('dark', 'dark');
+registerTheme('dark', getThemeByStylesheet(antvDark));
 
 // 注册 G 渲染引擎
 import * as CanvasEngine from '@antv/g-canvas';
