+++ src/theme/index.ts
@@ -1,7 +1,7 @@
-import { get, isString, lowerCase } from '@antv/util';
+import { get, lowerCase } from '@antv/util';
 import { LooseObject, StyleSheet } from '../interface';
 
-import { getThemeByStylesheet } from './default';
+import { getThemeByStylesheet } from './get-theme-by-style-sheet';
 import { antvLight as DefaultStyleSheet } from './style-sheet/light';
 
 const defaultTheme = getThemeByStylesheet(DefaultStyleSheet as StyleSheet);
@@ -24,33 +24,8 @@ export function getTheme(theme?: string): LooseObject {
  * @param theme 主题名。
  * @param value 具体的主题配置。
  */
-export function registerTheme(theme: string, value: LooseObject | string) {
-  if (isString(value)) {
-    // 在默认样式的基础上，修改主题样式表
-    const styleSheet = getStyleSheet(value);
-    Themes[lowerCase(theme)] = getThemeByStylesheet(styleSheet);
-  } else {
-    Themes[lowerCase(theme)] = value;
-  }
+export function registerTheme(theme: string, value: LooseObject) {
+  Themes[lowerCase(theme)] = value;
 }
 
-const StyleSheets: Record<string, StyleSheet> = {
-  default: DefaultStyleSheet,
-};
-
-/**
- * 获取对应主题样式表
- * @param name 样式表名
- */
-export function getStyleSheet(name?: string): StyleSheet {
-  return get(StyleSheets, lowerCase(name), StyleSheets.default);
-};
-
-/**
- * 注册主题样式表
- * @param name 样式表名
- * @param styleSheet 样式表定义
- */
-export function registerStyleSheet(name: string, styleSheet: StyleSheet) {
-  StyleSheets[lowerCase(name)] = styleSheet;
-};
+export { getThemeByStylesheet };
