+++ tests/unit/theme/index-spec.ts
@@ -1,4 +1,4 @@
-import { getStyleSheet, getTheme, registerStyleSheet, registerTheme } from '../../../src';
+import { getTheme, registerTheme } from '../../../src';
 
 describe('theme', () => {
   test('theme API', () => {
@@ -12,13 +12,5 @@ describe('theme', () => {
 
     expect(getTheme('a')).toEqual({ a: 1 });
     expect(getTheme('default')).toEqual({});
-
-    expect(getStyleSheet()).toEqual(getStyleSheet('default'));
-    expect(getStyleSheet('xxxxx')).toEqual(getStyleSheet('yyyyy'));
-    registerStyleSheet('default', {});
-    registerStyleSheet('a', { brandColor: 'red' });
-
-    expect(getStyleSheet('a')).toEqual({ brandColor: 'red' });
-    expect(getStyleSheet('default')).toEqual({});
   });
 });
